+++ pyproject.toml
@@ -2,3 +2,22 @@
 python_version = "3.10"
 ignore_missing_imports = true
 show_error_codes = true
+
+[build-system]
+requires = ["setuptools>=64"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "massconfigmerger"
+version = "0.1.0"
+description = "VPN subscription aggregator and merger"
+readme = "README.md"
+requires-python = ">=3.8"
+
+[project.scripts]
+aggregator-tool = "massconfigmerger.aggregator_tool:main"
+vpn-merger = "massconfigmerger.vpn_merger:main"
+vpn-retester = "massconfigmerger.vpn_retester:main"
+
+[tool.setuptools.packages.find]
+where = ["src"]
