+++ src/massconfigmerger/constants.py
