+++ tests/test_aggregator_clash_proxy.py
@@ -1,9 +1,6 @@
-import os
-import sys
 import yaml
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-import aggregator_tool
+from massconfigmerger import aggregator_tool
 from massconfigmerger.config import Settings
