+++ tests/test_clash_parsing_extra.py
@@ -1,9 +1,6 @@
-import os
-import sys
 import base64
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from clash_utils import config_to_clash_proxy
+from massconfigmerger.clash_utils import config_to_clash_proxy
 
 
 def test_ssr_parse_success():
