+++ tests/test_clash_proxies_yaml.py
@@ -1,11 +1,8 @@
-import os
-import sys
 import asyncio
 import yaml
 from pathlib import Path
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from vpn_merger import UltimateVPNMerger, ConfigResult, CONFIG
+from massconfigmerger.vpn_merger import UltimateVPNMerger, ConfigResult, CONFIG
 
 
 def test_clash_proxies_yaml(tmp_path, monkeypatch):
