+++ tests/test_deduplicate.py
@@ -1,10 +1,8 @@
 import base64
 import json
-import os
 import sys
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-import aggregator_tool
+from massconfigmerger import aggregator_tool
 from massconfigmerger.config import Settings
