+++ tests/test_fetch_text.py
@@ -1,13 +1,10 @@
-import os
-import sys
 import asyncio
 import pytest
 from aiohttp import web
 
 pytest_plugins = "aiohttp.pytest_plugin"
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-import aggregator_tool
+from massconfigmerger import aggregator_tool
 
 
 @pytest.mark.asyncio
