+++ tests/test_parse_configs.py
@@ -1,9 +1,6 @@
-import os
-import sys
 import base64
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-import aggregator_tool
+from massconfigmerger import aggregator_tool
 
 
 def test_multiple_links_same_line():
