+++ tests/test_processor_methods.py
@@ -1,11 +1,9 @@
 import base64
 import json
-import os
 import sys
 import types
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from vpn_merger import (
+from massconfigmerger.vpn_merger import (
     EnhancedConfigProcessor,
     ConfigResult,
     UltimateVPNMerger,
