+++ tests/test_scrape_telegram_configs.py
@@ -1,10 +1,7 @@
 import asyncio
 import types
-import os
-import sys
 
-sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-import aggregator_tool
+from massconfigmerger import aggregator_tool
 from massconfigmerger.config import Settings
